+++ deepxube/trainers/utils/train_loop.py
@@ -123,6 +123,7 @@ def train(domain: Domain, heur_nnet_par: Optional[HeurNNetPar], update_heur: Opt
                 updater.set_heur_file(heur_targ_file)
 
     if policy_nnet_par is not None:
+        update_policy.set_policy_samp(policy_samp)
         for updater in [update_heur, update_policy]:
             if (updater is not None) and isinstance(updater, UpdateHasPolicy):
                 updater.set_policy_nnet(policy_nnet_par)