+++ CHANGES.md
@@ -26,6 +26,7 @@
 * Add abstract method for loss info for training policy
 * Add layer norm to resnet_fc
 * Vectorize expand
+* Make policy at update_num=0 sampled from Domain's sample_state_action
 
 ## 0.2.1
 * Consolidate search: Beam search -> special cases: greedy_policy, graph search -> special cases: batch weighted A* search, batch weighted Q* search